+++ spec/open_meteo/client/config_spec.rb
@@ -15,12 +15,7 @@ RSpec.describe OpenMeteo::Client::Config do
 
       context "when the global config is set" do
         before { OpenMeteo.configure { |config| config.api_key = "global-api-key" } }
-
-        after do
-          OpenMeteo.configure do |config|
-            config.api_key = -> { ENV.fetch("OPEN_METEO_API_KEY", nil) }
-          end
-        end
+        after { OpenMeteo.configure { |config| config.api_key = nil } }
 
         it { is_expected.to eq "global-api-key" }
       end