+++ spec/open_meteo/integration/forecast/general_spec.rb
@@ -60,7 +60,9 @@ RSpec.describe "Integration > forecast > general" do
     context "when an API key is used" do
       subject(:response) { forecast.get(location:, variables:) }
 
-      let(:api_config) { OpenMeteo::Client::Config.new(api_key: "123-test", host: "customer-api.open-meteo.com") }
+      let(:api_config) do
+        OpenMeteo::Client::Config.new(api_key: "123-test", host: "customer-api.open-meteo.com")
+      end
       let(:client) { OpenMeteo::Client.new(api_config:) }
       let(:forecast) { OpenMeteo::Forecast.new(client:) }